move _enforce_nbf_exp to new module
